Rework

+++ Assets/Scripts/Queens/Models/Status.cs
@@ -5,9 +5,9 @@ namespace Queens.Models
     [Serializable]
     public class Status
     {
-        public float flow { get; set; }
-        public float popularity { get; set; }
-        public float health { get; set; }
-        public float money { get; set; }
+        public int flow { get; set; }
+        public int popularity { get; set; }
+        public int health { get; set; }
+        public int money { get; set; }
     }
 }
